Add ListWatchedAddresses RPC for Delegate Feature
This commit implements functionality for users to retrieve a list of
all addresses currently being monitored for VTXO rollovers. This gives
users visibility into which addresses the Ark node is watching as a
delegate.

The implementation:
- Adds ListWatchedAddresses RPC to the DelegateService
- Implements method to retrieve all rollover targets from repository
- Adds gRPC handler to expose the functionality

This is the fourth component of the delegate feature, providing users
with visibility into the delegation system.